a little bit of import cleanups
