--- src/Components/TaskManagerComponent/TaskManagerComponent/IOptionSetConfiguration.ts
@@ -1,6 +1,0 @@
-
-export interface IOptionSetConfiguration {
-    key: number;
-    label: string;
-    color: string;
-}
