--- src/Components/TaskManagerComponent/TaskManagerComponent/ITaskManagerBadgeConfigurationItem.ts
@@ -1,7 +1,0 @@
-import { IOptionSetConfiguration } from "./IOptionSetConfiguration";
-
-export interface ITaskManagerBadgeConfigurationItem {
-    name: string;
-    keys: number[];
-    values?: IOptionSetConfiguration[]
-}
