+++ src/Components/TaskManagerComponent/TaskManagerComponent/enums.ts
@@ -1,0 +1,10 @@
+export enum TaskStatus {
+    InProgress = 3,
+    Completed = 5,
+    Canceled = 6
+}
+
+export enum Culture {
+    en = 1033,
+    fr = 1036
+}
